refactor: rename cif_columns to cif_constants
